Refactor: 타입 정리, 워닝 해결
