+++ src/hooks/useSetOptions.ts
@@ -1,12 +1,5 @@
-import { Dispatch, SetStateAction, useEffect } from 'react';
-
-interface SetOptionsProps {
-  startDate: string;
-  lastDate: string;
-  currentYear: number;
-  currentMonth: number;
-  setCurrentMonth: Dispatch<SetStateAction<number>>;
-}
+import { SetOptionsProps } from '@/types/CalendarType';
+import { useEffect } from 'react';
 
 export default function useSetOptions({
   startDate,
