--- src/types/PostType.ts
@@ -1,12 +1,0 @@
-export interface PostType {
-  created_at: string;
-  d_day: number | null;
-  feeling_status: number;
-  goal: string | null;
-  id: number;
-  memo: string | null;
-  todo_date: string;
-  todo_progress: number;
-  updated_at: string;
-  user: number;
-}
